dhcp_sqlippools: Make initial address offers expire quickly
We want to make sure that the unaccepted offers are released quickly.

Especially true because during DISCOVER an unique address will be offered by
each server that is using the database, only one of which will ultimately be
requested. We don't want to fill the pool with long-lived, unused offers.