sqlippools: Support for requested IP address hint
Add req_attribute_name configuration item that specifies which request
attribute is to be used as a hint for the IP address to assign.

For RADIUS this is likely to be the Framed-IP-Address (IPv4) or
Framed-IPv6-Prefix (IPv6) attributes sent during Access-Request processing.

For DHCP this will be the DHCP-Requested-IP-Address parameter, usually sent in
renewals and during rebinding.

It is only a hint. It is passed to the new allocate_requested query (if
present) which will allocate the address if it is acceptable, according to
local policy.

Query updates for the various databases to follow...