Set a capacity for trunk backlog
And allow requests to be allocated / enqueued even if the max
connections and requests per connection limit is reached, up to the
backlog limit.

Helps handle spikes in traffic, particularly where trunks are limited to
one request per connection, and there are server side limitations on the
number of connections that can be made - e.g. SQL servers.