Storing redis result against the command doesn't work
When a multiple commands are sent as a command set, when the `demux`
callback is run, all commands are passed the same pointer to the result
object, with the object having been updated to be the result of that
command.

So, storing the result pointer does give access to anything beyond the
result of the last command.